enriching/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.65.0
pyyaml>=6.0

# Levenshtein distance (Wylie syllable-level)
python-Levenshtein>=0.21.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
